회원가입 된다! 로그인 되는것 같음! 12/30

+++ movie/src/main/java/movie/Servlet/user/LogoutServlet.java
@@ -19,32 +19,30 @@ public class LogoutServlet extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		// url : /users/idCheck
 		String root = request.getContextPath();
+		HttpSession session = request.getSession(false);
 		
-		// /logout - 로그아웃
-		System.out.println("로그아웃...");
-		HttpSession session = request.getSession();
+		Cookie rememberMeCookie = new Cookie("rememberMe","");
+		Cookie tokenCookie = new Cookie("token","");
+		rememberMeCookie.setPath("/");
+		tokenCookie.setPath("/");
+		rememberMeCookie.setMaxAge(0);
+		tokenCookie.setMaxAge(0);
 		
-		// 쿠키 제거
-		// - 자동 로그인, 토큰
-		Cookie rememberMeCookie = new Cookie("rememberMe", "");
-		Cookie tokenCookie = new Cookie("token", "");
-		Cookie[] deleteCookies = { rememberMeCookie, tokenCookie };
-		for (int i = 0; i < deleteCookies.length; i++) {
-			Cookie cookie = deleteCookies[i];
-			cookie.setPath(root);
-			cookie.setMaxAge(0);
-			response.addCookie(cookie);
+		response.addCookie(rememberMeCookie);
+		response.addCookie(tokenCookie);
+		
+		if(session != null) {
+			String username = (String) session.getAttribute("username");
+			if(username != null) {
+				PersistenceLoginsService persistenceLoginsService = new PersistenceLoginsServiceImpl();
+				persistenceLoginsService.delete(username);
+			}
+			session.invalidate();
 		}
-		
-		// 자동 로그인 토큰 삭제
-		PersistenceLoginsService persistenceLoginsService = new PersistenceLoginsServiceImpl();
-		String username = (String) session.getAttribute("username");
-		persistenceLoginsService.delete(username);
-		
-		session.invalidate();				// 세션 속성 모두 제거 : 세션 무효화
-		response.sendRedirect(root + "/");	// 메인 화면으로 이동
+		response.sendRedirect(root + "/");
 	}
 
+	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		doGet(request, response);
 	}
